tests

+++ src/erd.int.spec.ts
@@ -52,11 +52,9 @@ describe('erd', () => {
       const erd = new Erd({ prisma, knownas, });
       const graph = await erd._getActionsGraph();
 
-      console.log(graph);
-
       [
-        new RegExp('Person\\d+\\[' + fixtures.oswald.knownas + ']-->\\|' + fixtures.assassinated.name + '\\|Person\\d+\\[' + fixtures.jfk.knownas + '\\];', 'g'),
-        new RegExp('Person\\d+\\[' + fixtures.arthur.knownas + ']-->\\|' + fixtures.hosted.name + '\\|Person\\d+\\[' + fixtures.oswald.knownas + '\\];', 'g'),
+        new RegExp('Person\\d+\\[' + fixtures.oswald.knownas + ']-->\\|' + fixtures.assassinated.name + '\\|Person\\d+\\[' + fixtures.jfk.knownas + '\\]', 'g'),
+        new RegExp('Person\\d+\\[' + fixtures.arthur.knownas + ']-->\\|' + fixtures.hosted.name + '\\|Person\\d+\\[' + fixtures.oswald.knownas + '\\]', 'g'),
       ].forEach(re => {
         expect(graph).toMatch(re);
       });
@@ -80,7 +78,10 @@ describe('erd', () => {
       const exists = fs.existsSync(savepath);
 
       expect(exists).toBeTruthy();
-      fs.unlinkSync(savepath);
+
+      if (!process.env.CRUFT) {
+        fs.unlinkSync(savepath);
+      }
     });
 
   });
